% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borggroenen.r
\docType{data}
\name{borggroenen}
\alias{borggroenen}
\title{Dataset: Cross-tabulation of crime rate vs. 10 US states}
\format{
dataframe
}
\usage{
data(borggroenen)
}
\description{
Cross-tabulation (7x10) of crime rate across 10 US states.\cr
After: Borg-Groenen 2005 (Table 24.5).
}
\keyword{datasets}
