% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelland.r
\docType{data}
\name{yelland}
\alias{yelland}
\title{Dataset: Cross-tabulation of authors vs. alphabet letters}
\format{
dataframe
}
\usage{
data(yelland)
}
\description{
Cross-tabulation (15x16) of the frequency of alphabet letters across multiple authors.\cr
After: Yelland 2010 (page 3).
}
\keyword{datasets}
